relocate groupstack and update react-native
