+++ components/Groups.js
@@ -1,10 +1,6 @@
-import { createStackNavigator } from '@react-navigation/stack';
 import React from "react";
-import { View, Text, FlatList, Button, Alert, TouchableOpacity } from "react-native";
+import { View, Text, FlatList, Button, TouchableOpacity } from "react-native";
 
-import ViewGroup from './ViewGroup';
-import CreateEvent from './CreateEvent';
-import CreateGroup from './CreateGroup';
 import styles from './Styles';
 
 
@@ -97,27 +93,4 @@ const Groups = ({ navigation }) => {
     );
 }
 
-//This is the stack navigator for the groups tab so that viewgroup can be accessed by pressing on a group in the flatlist
-const Stack = createStackNavigator()
-
-const GroupStack = () => {
-    return (
-        <Stack.Navigator screenOptions={{ headerShown: false }}>
-            <Stack.Screen
-                name="Groups" component={Groups}
-            />
-            <Stack.Screen
-                name="ViewGroup" component={ViewGroup}
-            />
-            <Stack.Screen
-                name="CreateEvent" component={CreateEvent}
-            />
-            <Stack.Screen
-                name="CreateGroup" component={CreateGroup}
-            />
-        </Stack.Navigator>
-    )
-}
-
-
-export default GroupStack;
+export default Groups;
