+++ components/navigation/BottomStack.js
@@ -8,7 +8,7 @@ import * as SS from 'expo-secure-store'
 import LoginContext from '../LoginContext';
 import CustomHeader from '../CustomHeader';
 import HelloWorld from '../HelloWorld';
-import GroupStack from '../Groups';
+import GroupStack from './GroupStack';
 
 const LaunchPage = ({ navigation }) => {
     //placeholder waiting for groups to merge
